Removed binaries. Fixed missing else statement.
