+++ tools/echo_server/config.json
@@ -1,5 +1,5 @@
 {
-    "verbose": true,
+    "verbose": false,
     "logging": false,
     "secure-connection": false,
     "server-port": "9000",
